[General] Use floats instead of doubles for enchantmentCharge

+++ components/openmw-mp/Base/BaseEvent.hpp
@@ -12,7 +12,7 @@ namespace mwmp
         std::string refId;
         int count;
         int charge;
-        double enchantmentCharge;
+        float enchantmentCharge;
 
         int actionCount;
